Disable DesugarHtmlNodesPass in PassManager for JsSrc and run it in JsSrcMain.
This will eventually allow us to instrument the velog nodes for JsSrc backend only.

This CL should not affect the generated code; it does have a side effect that we run CombineConsecutiveRawTextNodesPass twice. We could add another options in the pass manager builder, but it might not worth it.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=165764168