Add an explicit pass for annotating the ast with print directives much like we do for soy functions.
Also, validate that print directives are called with the correct number of
arguments.

This will allow us to stop injecting the print directive map in so many places
and delete a number of binding.  As well as improving types in a number of
places.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=166384223