Fix a bug in StrictHtmlValidationPass that reports misleading errors.
The root cause is that we always add close tag to a queue if we cannot find a match in the current block. This is useful for case like

{if $foo}<div>{/if}
{if $foo}</div>{/if}

When we see the close tag in this case, we should not report an error immediately. Instead, we add these to a queue and try to match them later.

This causes issues when a close tag is presented and is outside of a control block. In this CL, I used a helper method that recursively find the parents and check if the current node is in a control block. An alternative will be to maintain a list of boolean during the tree traversal, but I guess the performance improvement is negligible and the code is less readable.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=167317115