Add SoyJbcSrcFunction and SoyJbcSrcPrintDirective interfaces
This will allow us to split up the monolithic PluginFunctionCompiler.  Also, by providing a print directive interface, we should be able to avoid the calling convention overheads in the most common cases.

For example {$foo|escapeHtml} currently compiles to

renderContext.getPrintDirective("escapeHtml").applyForJava(<autobox>$foo, ImmutableList.of());

in the future we should be able to change that to a simple

Sanitizers.escapeHtml(foo)

which is analagous to what the other backends do.  Follow up cls will migrate all the built-in functions and print directives to implement these interfaces.

For the time being these interfaces will not be available for plugin authors to use directly. This will be revisted in the future.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=168268476