Support for streaming print directives in jbcsrc
specifically:
* add a new jbcsrc specific interface for streaming print directives
* enhance the compiler to be able to apply the streaming directives and use
  them when all directives applied to an expression are streamable.
* modify the compiler to call the new appendable methods to advertise content
  kinds (the default implementations of these methods are no-ops.

The main print directives we want to support have special cases for certain
content kinds.  For example, |escapeHtml which is liberally used by the
autoescaper. This means we need to be able to communicate content kinds through
the appendable system (hence the pair of extra methods).  At this point there
are no non-test implementations of the Streamable interface so this should be a
no-op for current users.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=168708366