Rollforward [] which was rolled back in [] because it triggered double escaping in some templates.  The issue was in how we decided to enable no-op mode.  The previous logic was that we could switch to no-op mode only if the current content matched, the new logic is that we should be in no-op mode as long as we are underneath any matching content block.  I have also added a few tests to demonstrate the issue.
*** Original change description ***
Add support for streaming through |escapeHtml for jbcsrc

This has an interesting side effect of eliminating a type check.  Previously,
if we had something like

  {template .foo}
    {@param p : int}
    {$p}
  {/template}

That used to be compiled to something like

  appendable.append(Sanitizers.escapeHtml(((IntegerData) p.resolve()).coerceToString()));

Now it will be compiled to something like

  p.renderAndResolve(streamingEscapeHtml(appendable));

Note how the type check has been elided.  This is what caused the churn in
SoySauceTest, we were explicitly passing an object of the wrong type to trigger
a ClassCastException, but now it doesn't happen.  This isn't actually out of
spec (we have documented that soy's lazy evaluation model requires that some
type checks may never happen), and i don't see how it could cause failures.
The alternative would be to somehow have renderAndResolve perform the type
check, though im not sure it would be worth the effort.  I have added a test in
StreamingPrintTest to demonstrate this.

I believe with this cl log statements will survive the autoescaper in the most
common case.  I will continue to retro fit other escaping directives in this
way.

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=168906081