Remove uses of assisted-inject from soy by hand writing the factories
This actually revealed a number of circular dependencies between the factories
and some of the collaborators, which are now documented and should inform further refactorings.  The benefit of this change is that assisted injection requires generating dynamic implementations of the interfaces which is fairly complex and adds latency to injector creation.  It also is part of the overall effort to remove guice from compiler internals.

In a followup i will remove injection entirely from the Tofu and SoySauce
factories since they are relatively trivial. This should allow the compiler to
avoid loading classes entirely from backends that are not in use.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=169128367