Remove guice injection from the pysrc backend
With this change we should no longer load any of the pysrc code when we aren't
compiling for it. (customers have actually complained about this before)

This refactoring basically just handwrites the current wiring in PySrcMain.
This of course leaves in place the weird (preexisting) circular dependency. But
the circular dep is now completely obvious from the wiring code.  Followup cls
will do the same for incremental dom and jssrc, but i decided to start with
this since it is simpler and smaller.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=169468824