Rewrite embedCssInHtml to avoid java.util.regex.Pattern
embedCssInHtml is on the critical path of page rendering for many applications
since css is embedded in the <head> and even if the user passes us a safe css
object we still need to make sure it doesn't escape a <style> tag.  This
handwritten loop improves performance by a factor of about 13.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=169692406