Remove the OptimizeBidiCodeGenVisitor
This pass was meant to inline certain functions like bidiMark() when compiling
with a static BidiGlobalDir.  However, this is silly since the implementations
in bidiMark() et. al. already do this.  It appears that this was added to soy
before the plugin api was fully worked out so it maybe wasn't the case at the
time.  But as of now this is just a complex reimplementation of several builtin
plugins.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=169938557