Fix the tree copying routines so they can be performed with a single AST traversal.
Tree cloning is fairly straightforward but VarDefn objects cause problems.  Normally the ResolveNamesPass will ensure that all VarRefNodes get VarDefns. But when cloning trees some of the VarDefns get replaced and so they need to be updated again.  This is currently handled by cloning the tree and then traversing over it again to fix the vardefns if they have changed.  The new solution is to do this iteratively during cloning which will allow us to get rid of the second passes.

To actually do this i have implemented a simple callback system so nodes can register to listen to changes to certain objects and when new VarDefns are created the listeners can be notified.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=170276923