Fix a bug where jbcsrc would apply print directives in the opposite order if there were multiple streaming directives.
This is unlikely to have affected many applications so far because there are
relatively few streaming print directives so far and of the ones we have they
are generally only applied by the autoescaper and when the autoescaper applies
multiple directives it doesn't (afaict) apply two streaming ones.

So a user would have had to manually apply some directives (as demonstrated in
the integration test).

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=170509934