Add support for closeable print directives in jbcsrc
This will be useful for a number of print directives that are
difficult/impossible to implement without buffering content at least
temporarily.

Rather than adding a close() method to LoggingAdvisingAppendable and always
calling it (And propagating it to delegates), i made it so that an appendable
could optionally implement closeable and the compiler would make sure it would
get called.  This complicates the code generator slightly, but makes it so that
delegating implementations don't ever need to worry about calling .close().

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=170537131