Remove the gensoypy.py binary wrapper script
This mostly just provided alternative flag names for all the compiler flags, the one useful thing it did was to calculate the --inputPrefix flag.  Which was mostly useful as a way to not have to specify a path for the manifest file to write out.

Instead i just modified the compiler to take an explicit path for the manifest file to write.  This is slightly more verbose on the command line, but IMHO less confusing than the way things were working before.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=171247063