Soy: plumb experimental features through to SoyFileParser.
SoyGeneralOptions has a nice getExperimentalFeatures() API that can be used to allow new language features in the Soy codebase (i.e., integrationtests) without releasing it generally. I would like to use it to introduce an experimental type.

This CL plumbs the SoyGeneralOptions available in SoyFileSet through to a new (optional) field in SoyFileSetParser.

It would be nice to then plumb the general options through to SoyFileParser, but this is infeasible. (SoyGeneralOptions has a dep on SoyUtils for parsing globals files, and SoyUtils in turn depends on the parser, so adding the general options to the parser creates a cyclic dep. We could eliminate the SoyGeneralOptions->SoyUtils dep by inlining the global file parsing methods into callers, but unfortunately there are ~10 callers throughout [], so it's probably not worth the effort.)

This CL therefore just extracts the set of experimental features from SoyGeneralOptions and passes it into SoyFileParser.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=175346172