Soy: add experimental_map type.
[] describes a migration during which Soy templates across [] will use both `legacy_object_map` and `map` keywords. Currently, the two keywords are aliases. We are going to migrate everyone from `map` to `legacy_object_map`. Then we can ban `map` and change its behavior in the type system.

This makes our progress on the underlying type system depend on the migration: we have to move everyone off `map` before we can change it in the type system. (We can, and have, introduced a new map type, MapType, that cannot be instantiated from Soy templates, but it is unclear how we are going to test it if we can't write Soy templates that use it.)

This CL introduces a third map keyword: `experimental_map`. It is connected to MapType, makes use of Soy's experimental features API so that only our integration tests can use it. Once the new type is mature and everyone is migrated to `legacy_object_map`, we can connect `map` to MapType, change all `experimental_map`s in our integrationtests to `map`s, and delete `experimental_map`.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=175616050