Soy: rename MapLiteralNode to LegacyObjectMapLiteralNode.
This is prefactoring for changing the Soy parser to support a new map literal syntax (and consequently, a new AST node).

The node class is a compiler implementation detail, so renaming it is safe. (Unlike the runtime MapType, which is used in plugin implementations.)

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=175994512