Soy: basic support for experimental_map in all backends.
Now that syntax exists for creating values of type experimental_map (map(k1: v1, k2:v2, ...)), we can start testing experimental_maps in Soy's integration tests.

This CL changes the integration tests to exercise experimental_maps wherever legacy_object_maps are currently being exercised, and makes the minimal changes to all four backends to get the tests passing. (The backend changes are conceptually independent, and could be sharded into separate CLs. I put them in one CL because the Java and Python approaches are similar, and because the primary test covering them is the cross-language integration test.)

experimental_map and legacy_object_map are incompatible at runtime in the JS backend, and because of Soy's guarantee of identical behavior across backends, they must be incompatible in the other backends too. This CL takes the shortcut of *breaking* that guarantee: since the Python and Java implementations of experimental_map re-use the existing implementations of legacy_object_map, it is wrongly possible for them to interoperate at runtime.

This CL adds TODOs in the Java and Python backends documenting what must be done in order to re-establish the cross-language behavior guarantee. I believe that temporarily violating it, and then repairing it, will allow us to land this feature more quickly by improving its integration test coverage. (And usage of experimental_map and map() literals is limited to Soy's codebase, so this should not break anyone or cause people to depend on this inconsistent behavior.)

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=177042920