Roll forward [] (Soy: add legacyObjectMapToMap() and mapToLegacyObjectMap() conversion plugins) with fix.
The OCL was rolled back for the first time because it broke an ads conformance test. I fixed the test in [] and rolled forward.

The OCL was rolled back for the second time because it broke a [] size test. I fixed the test in []

This CL rolls forward the OCL for the third time.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=177841386