Roll forward [] (Soy: add legacyObjectMapToMap() and mapToLegacyObjectMap() conversion plugins) for the 4th time.
Details of previous rollbacks/rollforwards can be found in the individual CLs:

- OCL:             []
- 1st rollback:    []
- 2nd rollforward: []
- 2nd rollback:    []
- 3rd rollforward: []
- 3rd rollback:    []

The most recent rollback, [] was caused by an IE-only integration test failure. The test loads a bundle of raw JS. Although the new code is never executed, it uses a for...of loop, which is a parse error in IE. (While modern JS integration tests are automatically transpiled to JS appropriate for each browser, the failing test is a legacy Puppet test that has no transpilation support.)

This 4th rollforward fixes the test to load a special bundle that does not include the for...of loop. (The Soy codebase internally needs this bundle for other reasons.)

Note that this rollforward also contains the changed introduced in the 3rd rollforward, since those changes cannot be submitted separately from the Soy changes (see discussion on []

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=178377546