Add Closure "disable debug" flags so that the compiled JS tests run closer to how they run in production.
One affect this has is that it disables assertions, which allows us to start running tests with assertions in them.

Also, add a define to determine whether to include the soyTemplateName field on each template. The soyTemplateName field is required to show the test names in the compiled tests, but would get stripped out because goog.DEBUG is now false. This new define defaults to goog.DEBUG so should be the same value as it was before, except where we now override it in the integration tests so with goog.DEBUG set to false we still include the soyTemplateName field.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=178924934