Add a Specialization of foreach $i in range(...) in TOFU
I plan on doing this in all backends, but tofu is generally the simplest, so i am starting there.

The benefit of the specialization is that we can avoid allocating the list object altogether.  This does require some redundancy, but less overall than the previous approach of using the for loop.

I also added a simple benchmark since the only expectation is improved performance.  This shows that:

with this optimization: 120224 bytes allocated, 3998 objects allocated, 277 microseconds
without this optimization: 123072 bytes allocated, 4006 objects allocated, 277 microseconds

so a pretty minor improvement.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=179221760