Add a Specialization of foreach $i in range(...) in jssrc
This is purely an optimization, but it is an enabling optimization since it will make it possible for people to migrate from for->foreach with no regression in performance.

One thing to note is that (afaict) the jscompiler has no current support for
constant folding across Math.* functions.  So prior to initializing a migration
i am going to investigate adding some peephole optimizations to the jscompiler
to support these functions.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=179435328