Automated [] rollback of changelist 180566896.
*** Reason for rollback ***

causing an autoescaping error

*** Original change description ***

When extracting message variables, don't do it if it is already wrapped in a variable.

In addition to being a small optimization, this will also be necessary for the text auditing work which expects to find msg nodes wrapped in let blocks, but that analysis breaks when they are wrapped in an additional layer of lets!

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=180577943