fix a performance problem with msg bundle lookup.
Constructing the SoyMsgBundleWithFulllocale object is expensive, so we should cache that instead of just the original bundle.  The reason it is expensive is due to the icu libraries we are using, calculating language tags and testing for RTL both take a long time.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=180840168