Delete the ExtractMsgVariablesVisitor and instead rely on the CodeChunk api.
The ExtractMsgVariablesVisitor is solving a problem that we don't really have.   It was designed so that goog.getMsg calls would be easy to generate as statements, but the CodeChunk api already solves this problem for us.  Additionally, this pass is also the slowest part of code generation (takes about 70% of the time of JsSrcMain because it is spends a lot of time calculating where the lets should be generated).

This required some non-trivial refactoring of the msg generation code to be more codechunky.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=183323237