add text auditing support to soy
This cl introduces a new builtin function called msgId(..) that can be used to access the id of a soy msg.  This works much like the index(),isLast(),isFirst() functions where the function must take a variable as an argument and the variable must be a let node that contains a single message.

The implementation consists of the following:
* 2 new builtin function definitions: msgId and $$isPrimaryMsgInUse (and implementation detail)
* MsgIdFunctionPass - a new SoyFilePass that validated msgId calls and handles all calls that do not involve a fallbackmsg, for those that do involve a fallback it rewrites the function call to $$isPrimaryMsgInUse($var, NNNN, MMMM) ? 'NNNNN' : 'MMMMM'.
* support in all 4 backends for the $$isPrimaryMsgInUse function call.  For all backends besides jssrc this was trivial, for jssrc it require a bit more trickiness due to how goog.getFallbackMsg works.

See []

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=183330679