Automated [] rollback of changelist 183330679.
*** Reason for rollback ***

Needed to cleanly rollback [] which caused b/72536588 and b/72534909

*** Original change description ***

add text auditing support to soy

This cl introduces a new builtin function called msgId(..) that can be used to access the id of a soy msg.  This works much like the index(),isLast(),isFirst() functions where the function must take a variable as an argument and the variable must be a let node that contains a single message.

The implementation consists of the following:
* 2 new builtin function definitions: msgId and $$isPrimaryMsgInUse (and implementation detail)
* MsgIdFunctionPass - a new So...

***

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=183386323