Automated [] rollback of changelist 183323237.
*** Reason for rollback ***

Needed to cleanly rollback [] which caused b/72536588 and b/72534909

See comments in [] for more context.

Calendar [] failure sponges for b/72534909 are:
[](@183314556) Passes
[](@183314557) Fails consistently

Tasks has verified the build interactively before and after the CL.

*** Original change description ***

Delete the ExtractMsgVariablesVisitor and instead rely on the CodeChunk api.

The ExtractMsgVariablesVisitor is solving a problem that we don't really have.   It was designed so that goog.getMsg calls would be easy to generate as statements, but the CodeChunk api already solves this problem for us.  Additionally, this pass is also the slowest part of code generation (takes about 70% of the time of JsSrcMain because it is spends a lot of time calculating where the lets should be generated).

This...

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=183392351