Rollforward [] which was rolled back in cl/ because it broke an incremental dom template which passed a let with kind="text" as a key attribute value.  The cl to introduce the autoescaper would cause such attribute values to be represented as UnsanitizedText objects.  This rollforward fixes the bug by implicitly coercing the value to a string.
*** Original change description ***

Start running the autoescaper for incrementaldom.

Due to how incrementaldom works many of the most common escaping directives are
unnecessary and actually harmful.  For example, |escapeHtml will introduce
double escaping since the incrementaldom.text function will use dom Text node
APIs to manipulate textual content.  This CL changes the compiler to
unconditionally run the autoescaper but then adds an incrementaldomsrc specific
compiler pass to remove a whitelisted subset of the directives added by the
autoescaper.  This way new directives will get automatically added and every
exemption will be opt-in.

In addition to running autoescaper this now requires that let and param blocks be wrapped in SanitizedContent objects so that they are handled correctly by the autoescaper runtime.

Of course this is not the ideal solution since it isn't obvious that this is
safe and it requires examining the incrementaldom library to be sure.
Additionally, there are other issues in the incrementaldom backend that will
require more work (like how certain expressions are handled in attributes
context).  And finally, because incrementaldom doesn't run many of the
integration tests it isn't obvious how safe a change like this is.

***

GITHUB_BREAKING_CHANGES=some incrementaldom templates will now have their print directives escaped.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=183456453