Refactor deltemplate variant validation to no longer rely on SoySyntaxException and delete some newly dead features from LegacyInternalSoySyntaxException.
deltemplate variant validation is pretty complex.  The rules are:

* if the variant is a string, it must be an identifier
* if the variant is an integer, it must be non-negative
* the variant can be no other type.

however, the most common use is a proto enum or a global reference of some other kind, so we can't always validate during construction.  To fix this i added a simple api to GlobalNode to plug into the resolution process.  Another option would be to explicitly validate during RewriteGlobalsPass

Finally, this resolves a bunch  of duplication of logic between TemplateDelegateNode and TemplateDelegateNodeBuilder.  Things were trying to be clever for when the variants were string literals, but this just made it unclear what was responsible for validation.  Sadly this CL doesn't resolve all the WTFs with this code.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=184307473