Add support for plural/select messages to contain ICU metacharacters
This was banned in the original support for plural/select because the original author thought that the escaping syntax was 'weird' and didn't want to support it.  The syntax has been stable for a long time so i think it is fine to support it.  The main motivation for this however is that it allows us to delete LegacyInternalSoySyntaxException.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=184944533