Override actualCustomStringRepresentation() instead of getDisplaySubject().
The intention (from []is:
- getDisplaySubject() is the method that people call.
- actualCustomStringRepresentation() is the method that people override.

(Well, actually, getDisplaySubject() is deprecated, so people should call actualAsString(). We'll take care of that eventually, too :))

getDisplaySubject() returns almost the same thing as actualCustomStringRepresentation(), except:

1. It puts <...> around it.
2. It automatically prepends any name set by a call to named(...).

So, by changing these subjects to override actualCustomStringRepresentation() instead, we save them from the trouble of doing those 2 things.

Naturally, some of the implementations don't current do those 2 things, and it's not clear whether they intended this or not. For example:
http://[]
http://[]
http://[]

Luckily, I don't see any evidence of those subjects being used with named(), so that part of the change shouldn't make a difference in practice. However, by switching which method we override, we will be adding <...> in the error message. This is unfortunate, but it should be temporary, as we're likely to remove <...> from Truth's failure messages altogether -- because, as these people have apparently noticed, it usually makes things harder to read. But again, one thing at a time :)

Tested:
    [] --sample for global presubmit queue
    []

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=185554507