Soy Java API: breaking change: change the meaning of SoyMap.
This CL is analogous to [] which changed the meaning of the `map` type in the Soy language. The current CL makes a similar renaming, but to the Java API instead of the Soy language.

tl;dr:

|-----------------------------------------------------------------------------|
| Soy language           | Java API type at head  | Java API type in this CL  |
|------------------------|----------------------------------------------------|
| map<K,V>               | SoyNewMap              | SoyMap                    |
| legacy_object_map<K,V> | SoyMap                 | SoyLegacyObjectMap        |
|-----------------------------------------------------------------------------|

In words:

Currently, SoyMap is the Java representation of the `legacy_object_map` Soy type. We want to represent the `map` type instead. To do this, [] introduced SoyLegacyObjectMap as a vacuous subinterface of SoyMap. I then incrementally migrated all Google users of SoyMap to SoyLegacyObjectMap. Since SoyMap is now unused, it is safe to change its meaning. Accordingly, this CL renames SoyNewMap to SoyMap.

GITHUB_BREAKING_CHANGES=You must change all references to SoyMap in Java code that calls Soy or implements plugins to SoyLegacyObjectMap.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=185571432