Correctly pack and unpack Safe*Proto types from proto maps.
When a Safe*Proto is taken out of a proto map it needs to be unpacked to a Soy SanitizedContent object. When a Soy SanitizedContent object is put into a proto map it needs to be packed into a Safe*Proto.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=186633112