Rollforward of []Correctly pack and unpack Safe*Proto types from proto maps.
When a Safe*Proto is taken out of a proto map it needs to be unpacked to a Soy SanitizedContent object. When a Soy SanitizedContent object is put into a proto map it needs to be packed into a Safe*Proto.

NEW: Evaluate ProtoCall parameters when creating the ProtoCall instead of when using it. This is required because the JS backend compares the ProtoCall of each member of a union type to the others (http://[]/third_party/java_src/soy/java/com/google/template/soy/jssrc/internal/TranslateExprNodeVisitor.java?l=500&rcl=186641254), to ensure it knows if/how to unpack a value. The original CL was rolled back because the FieldDescriptor was stored in the ProtoCall, but FieldDescriptors don't have to actually be equal to know if/how to unpack a value (i.e. the parent message type can be different). Now, this only stores the details it needs in the ProtoCall so that it can correctly compare ProtoCalls in union types.

OPEN_SOURCE_BREAKING_CHANGES=none

Automated [] rollback of changelist 186641254.

*** Reason for rollback ***

Rollforward, fixing the issue that broke union types.

*** Original change description ***

Automated [] rollback of changelist 186633112.

*** Reason for rollback ***

[] has detected that 500 or more targets failed to build at [] If this is an error, please file a bug at []
[] double-checked a few of the targets built at []
[]
but failed to build at [] Sponge link for all broken targets:
[]

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=187190460