Delete SoyFileInfo.getProtoTypes(), the getProtoDescriptors() should be used instead.
The getProtoTypes() method had an inconvenient return type (ImmutableList<Object>) because at the time we wanted to avoid a dependency on proto in the opensource build.  That issue has since been fixed and so now it is time to fix the return type.  This has been done by adding a new method getProtoDescriptors() which returns the exact same data but with a proper type.

GITHUB_BREAKING_CHANGES=users of SoyFileInfo.getProtoTypes() should switch to calling the getProtoDescriptors() method.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=187358730