Rewrite a bunch of jssrc tests.... but not all of them
* turn a bunch of the tests into either more fine grained unit tests (like the type assertion tests) or integration tests
* remove tests that are testing error conditions which are covered by the compilation tests.

the motivation is that maintaining these tests is very tedious and because they hook into the GenJsCodeVisitor in weird ways it makes it difficult to modify some parts of the jssrc backend.  This cl doesn't get all of the tests, but a fair number of them.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=187390508