Allow xid() to take a global parameter
This automatically makes xid work with alias which is the main motivation.  It
also will allow for slightly shorter xid calls since they won't need to add the
quotation characters. One potentially controversial thing is that i made it an
error for the global to actually resolve to anything (either an explicit global
value or a proto enum).  This isn't necessary for correctness but it does make
it unclear what the author intention was, so i just banned it.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=187650301