Soy: change parameter types of isNull, isNonnull from `?` to `any`.
In order for the JS backend to generate correct code, the `map` type cannot be assigned to or from the unknown type (see xrefed bug). Meanwhile, the isNull and isNonnull functions require ?-typed params. When I tried to migrate a ?-typed param that happened to be passed to isNonnull in [] it caused a compilation error: []

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=189665944