Soy: get SoyMapData working with `map` values.
The recommended way to render a Soy template in Java is to pass the Renderer a java.util.Map<String, ?> of parameter names to parameter values. The values are idiomatic Java values, e.g. a Soy `list` param can be rendered with an ArrayList. The values are converted to internal Soy data structures automatically. This code path already works with the new `map` type.

There is an older rendering API in which Java callers can directly construct the internal data structures that back Soy parameters before passing them into the Renderer. There is no reason to use this API, since it involves classes that have little value elsewhere in an application. But the old API has many users, and we have never migrated them to the new API. So we need to make the old API work with the new `map` type.

The old API class for representing legacy_object_map/record values is SoyMapData. This CL changes SoyMapData to also implement SoyMap (the new map type), and equips it with the same runtime type detection logic recently added to DictImpl (for the new API code path). The runtime type detection logic is extracted to a collaborator, RuntimeMapTypeTracker.

Finally, this CL marks SoyMapData as @Deprecated.

GITHUB_BREAKING_CHANGES=no breakages, but marked SoyMapData as @Deprecated. There is no reason to use it. Use normal java.util.Maps instead.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=190126798