SoyMapData: remove RuntimeMapTypeTracker logic.
The purpose of RuntimeMapTypeTracker is to detect conflicting calls to map/legacy_object_map accessors within the same rendering call (e.g. rendering a SoyMapData as a map in one template and a legacy object map in a callee) and throw a runtime exception. Such conflicting calls always cause a runtime error in JS, so it is valuable for the Java backends to have similar behavior.

RuntimeMapTypeTracker works correctly as a collaborator with DictImpl (one of the runtime map classes). DictImpl instances are always created by the Soy runtime at the beginning of each render call. By contrast, SoyMapData objects are created and owned by callers and used for potentially multiple template renders. If a SoyMapData object is used as a map in one render call and a legacy object map in a subsequent call, RuntimeMapTypeTracker will incorrectly throw an exception.

This CL removes the RuntimeMapTypeTracker logic from SoyMapData. This may cause some Java renders to succeed where their corresponding JS renders fail, but it is a temporary state designed to ease the global migration from legacy object maps to maps.

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=190247651