Remove the --inputPrefix flag from Soy
This is being removed because it is kind of confusing (only applies to a subset of inputs), not really necessary and underused.

GITHUB_BREAKING_CHANGES=uses of the inputPrefix flag will need to switch to fully qualifying the soy file names.  A shell alias should be a simple alternative.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=191653259