Remove SyntaxVersion inference from the Soy compiler as well as everything except the 1.0 and 2.0 versions.
With this cl syntax version is controlled solely by the --shouldAllowDeprecatedSyntax flag in the jssrc backend.  In all other backends it is always 2.0 (or 1.0 in the case of the msg extraction/pruning rules).  This CL also removes a bunch of references to SyntaxVersion where it is no longer relevant.

GITHUB_BREAKING_CHANGES=Remove the --syntaxVersion version flags from the jssrc and pysrc compilers. Instead the --shouldAllowDeprecatedSyntax can be used in jssrc to enable deprecatedV1 templates. (The recommended workaround is to migrate your templates to modern Soy language features.)

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=191681365