Remove the ability to configure soy to only provide templates and not namespaces.
Going forward the behavior of the soy compiler will be as if --shouldProvideSoyNamespaces and --shouldProvideBothSoyNamespacesAndJsFunction have been set and the actual  flags are now ignored.  This is being done in order to eventually enable a migration to goog.module.  Soon after this cl we will stop generating provides for individual templates and instead users should solely rely on the provides for the namespaces.  This is an intermediate step in that process.

By doing this we can avoid the complexities of coordinating changes in these flags across multiple independent projects, instead all invocations of the compiler will be implicitly compatible.

GITHUB_BREAKING_CHANGES=The goog.provide strategy for the jssrc backend is changing and configuration APIs are being removed.  This should be a compatible change as long as the Soy namespaces are unique.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=191768696