Add a few micro optimizations to jbcsrc.
When rendering a SoyValueProvider incrementally, only stash the instance in the
currentRenderee field if it is non 'cheap'.  A common case is to render a
parameter which is always cheap to access.

When invoking a JbcSrc plugin we need to hand it the JbcSrcPluginContext
object, this allows plugins to access things like the current locale.  However,
most plugins don't care.  The implementation the RenderContext parameter
field, however a side effect of accessing RenderContext inside a let or param
is that we generate code to copy it into the new inner class (add a constructor
param, a field and code to assign it in the constructor).  If the
object isn't actually needed, this is a waste.  So instead return a lazy
delegate.

These two changes should reduce the size of lazy closures and the number of fields allocated for templates.

These changes revealed that DetachState.detachForRender had the implicit dependency that the callRender Expression did not rely on any local variables.  Since this change invalidates that assumption a new (simpler) detachForRender method has been added and the old one was renamed to detachForCall

GITHUB_BREAKING_CHANGES=none

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=193970121