Automated [] rollback of changelist 193970121.
*** Reason for rollback ***

Breaking []

Sample failure: []
*** Original change description ***

Add a few micro optimizations to jbcsrc.

When rendering a SoyValueProvider incrementally, only stash the instance in the
currentRenderee field if it is non 'cheap'.  A common case is to render a
parameter which is always cheap to access.

When invoking a JbcSrc plugin we need to hand it the JbcSrcPluginContext
object, this allows plugins to access things like the current locale.  However,
most plugins don't care.  The implementation the RenderContext parameter
field, however a side effect of ac...

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=193980262